+++ src/FormFields.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Ycs77\LaravelFormFieldType;
+
+abstract class FormFields
+{
+    /**
+     * Return form fields array.
+     *
+     * @return array
+     */
+    abstract public function fields();
+}